ci: fix GitHub Actions, migrate to rustls, bump to v0.2.0
The CI badge had been red on every run since the repo went public. Repair
all four failing jobs:

- Clippy (-D warnings): apply clippy --fix style lints across modules and add
  scoped #[allow(dead_code)] where the binary re-compiles library modules.
- Security Audit: update the moved action reference to rustsec/audit-check@v2.
- Linux musl: migrate reqwest from native-tls to rustls-tls, removing the
  openssl-sys dependency that broke the static build; add cmake for aws-lc-rs.
- Windows: install the Npcap SDK so pnet can link Packet.lib.

Dropping OpenSSL also eases static / musl / Alpine builds and shrinks the
native attack surface. Bump version to 0.2.0.

+++ src/modules/passive/llmnr.rs
@@ -215,7 +215,7 @@ fn decode_nbt_name(encoded: &[u8]) -> String {
         if chunk.len() < 2 {
             break;
         }
-        let c = (((chunk[0] as u8 - b'A') << 4) | (chunk[1] as u8 - b'A')) as char;
+        let c = (((chunk[0] - b'A') << 4) | (chunk[1] - b'A')) as char;
         if c.is_ascii() {
             name.push(c);
         }
